requirements.txt
aiolimiter==1.2.1
basicsr==1.4.2
cachetools==5.5.2
diskcache==5.6.3
fastapi==0.115.14
langchain_core==0.3.66
langchain_openai==0.3.26
matplotlib==3.10.3
numpy==2.3.1
opencv_python==4.10.0.84
Pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
sahi==0.11.26
scipy==1.16.0
streamlit==1.44.1
streamlit_extras==0.7.1
streamlit_shortcuts==0.1.9
tenacity==9.1.2
uvloop==0.21.0
